backend/requirements.txt
Django>=5.1,<6.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.3,<5.0
django-filter>=24.0
django-extensions>=3.2
psycopg2-binary>=2.9,<3.0
pgvector>=0.3,<1.0
gunicorn>=22.0,<23.0
redis>=5.0,<6.0
django-redis>=5.4,<6.0
Pillow>=10.0,<11.0
python-dotenv>=1.0,<2.0
openai>=1.30,<2.0
whitenoise>=6.5,<7.0
drf-spectacular>=0.27
markdown>=3.5
